Move form view models to forms folder
