+++ Frontend/Frontend.csproj
@@ -29,4 +29,7 @@
     <PackageReference Include="Microsoft.AspNetCore.DataProtection.StackExchangeRedis" Version="5.0.3" />
     <PackageReference Include="Microsoft.Extensions.Caching.StackExchangeRedis" Version="5.0.1" />
   </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Models\Projects" />
+  </ItemGroup>
 </Project>
