+++ Frontend/Models/FeaturesViewModel.cs
@@ -2,6 +2,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Data.Models.Projects;
 using Frontend.Helpers;
+using Frontend.Models.Forms;
 
 namespace Frontend.Models
 {
