+++ Frontend/Models/Forms/DateInputViewModel.cs
@@ -1,0 +1,9 @@
+namespace Frontend.Models.Forms
+{
+    public class DateInputViewModel
+    {
+        public string Day { get; set; }
+        public string Month { get; set; }
+        public string Year { get; set; }
+    }
+}
