+++ Frontend/Models/Forms/FormErrorsViewModel.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using System.Linq;
 
-namespace Frontend.Models
+namespace Frontend.Models.Forms
 {
     public class FormErrorsViewModel
     {
