+++ Frontend/Views/Shared/_FormErrorSummary.cshtml
@@ -1,4 +1,4 @@
-@model FormErrorsViewModel
+@model Frontend.Models.Forms.FormErrorsViewModel
 @if (Model.HasErrors)
 {
     <div class="govuk-grid-row">
