+++ Frontend/Views/Shared/_InlineRadioButtons.cshtml
@@ -1,4 +1,4 @@
-@model List<RadioButtonViewModel>
+@model List<Frontend.Models.Forms.RadioButtonViewModel>
 
 <div class="govuk-radios govuk-radios--inline">
     @foreach (var radioButton in Model)
