+++ Frontend/Views/Shared/_RadioButtons.cshtml
@@ -1,4 +1,4 @@
-@model List<RadioButtonViewModel>
+@model List<Frontend.Models.Forms.RadioButtonViewModel>
 
 <div class="govuk-radios">
     @foreach (var radioButton in Model)
